+++ src/uni_modules/fant-mini-plus/components/hd-area/types.ts
@@ -1,8 +1,8 @@
 /*
- * @Author: 徐庆凯
+ * @Author: weisheng
  * @Date: 2022-12-14 19:34:38
  * @LastEditTime: 2023-03-21 21:36:41
- * @LastEditors: 徐庆凯
+ * @LastEditors: weisheng
  * @Description:
  * @FilePath: \fant-mini-plus\src\uni_modules\fant-mini\components\hd-area\types.ts
  * 记得注释