+++ src/uni_modules/fant-mini-plus/components/hd-empty/hd-empty.vue
@@ -1,8 +1,8 @@
 <!--
- * @Author: 徐庆凯
+ * @Author: weisheng
  * @Date: 2021-12-30 21:28:28
  * @LastEditTime: 2023-03-23 17:01:31
- * @LastEditors: 徐庆凯
+ * @LastEditors: weisheng
  * @Description: 
  * @FilePath: \fant-mini-plus\src\uni_modules\fant-mini\components\hd-empty\hd-empty.vue
  * 记得注释