+++ src/uni_modules/fant-mini-plus/components/hd-loading/index.ts
@@ -1,8 +1,8 @@
 /*
- * @Author: 徐庆凯
+ * @Author: weisheng
  * @Date: 2022-12-14 17:33:21
  * @LastEditTime: 2023-03-22 14:58:48
- * @LastEditors: 徐庆凯
+ * @LastEditors: weisheng
  * @Description:
  * @FilePath: \fant-mini-plus\src\uni_modules\fant-mini\components\hd-loading\index.ts
  * 记得注释