+++ src/uni_modules/fant-mini-plus/components/hd-loading/types.ts
@@ -1,8 +1,8 @@
 /*
- * @Author: 徐庆凯
+ * @Author: weisheng
  * @Date: 2022-12-15 11:45:30
  * @LastEditTime: 2023-03-21 21:49:24
- * @LastEditors: 徐庆凯
+ * @LastEditors: weisheng
  * @Description:
  * @FilePath: \fant-mini-plus\src\uni_modules\fant-mini\components\hd-loading\types.ts
  * 记得注释