+++ src/uni_modules/fant-mini-plus/components/hd-swipe-action/index.ts
@@ -1,8 +1,8 @@
 /*
- * @Author: 徐庆凯
+ * @Author: weisheng
  * @Date: 2023-03-21 19:57:00
  * @LastEditTime: 2023-03-22 12:50:35
- * @LastEditors: 徐庆凯
+ * @LastEditors: weisheng
  * @Description:
  * @FilePath: \fant-mini-plus\src\uni_modules\fant-mini\components\hd-swipe-action\index.ts
  * 记得注释