+++ src/uni_modules/fant-mini-plus/components/hd-tab/hd-tab.vue
@@ -1,8 +1,8 @@
 <!--
- * @Author: 徐庆凯
+ * @Author: weisheng
  * @Date: 2021-10-14 19:28:57
  * @LastEditTime: 2023-03-27 14:10:15
- * @LastEditors: 徐庆凯
+ * @LastEditors: weisheng
  * @Description: 
  * @FilePath: \fant-mini-plus\src\uni_modules\fant-mini\components\hd-tab\hd-tab.vue
  * 记得注释