+++ src/uni_modules/fant-mini-plus/libs/utils/DatePicker.ts
@@ -1,8 +1,8 @@
 /*
- * @Author: 徐庆凯
+ * @Author: weisheng
  * @Date: 2022-12-15 19:43:51
  * @LastEditTime: 2022-12-16 17:42:12
- * @LastEditors: 徐庆凯
+ * @LastEditors: weisheng
  * @Description:
  * @FilePath: \fant-mini\src\uni_modules\fant-mini\libs\utils\DatePicker.ts
  * 记得注释